feat: refactor click and focus utilities to use a shared waitForThemed function for improved code reuse
